utils/pyproject.toml
[project]
name = "eit-cody-authx"
version = "0.1.0"
description = "JWT Generator für Tests"
requires-python = ">=3.12"
dependencies = [
    "pyjwt[crypto]>=2.8",
    "pyyaml>=6.0",
]

[tool.uv]
managed = true
